refactor(dtx): update barrier handling and improve gRPC integration
- Refactored barrier.go to utilize the built-in BarrierFromGrpc function from the dtmgrpc package.
- Updated error handling in dtm.go to check for gRPC server configuration instead of client initialization.
- Enhanced example.go by removing the TCC transaction submission error handling.
- Adjusted utils.go to improve gRPC server configuration checks.
- Updated go.mod files across multiple plugins to include new dependencies and versions.
- Removed unused files and improved overall code structure for better maintainability.

+++ plugins/mq/rocketmq/health.go
@@ -150,10 +150,10 @@ func (hc *HealthChecker) GetLastCheck() time.Time {
 }
 
 // GetErrorCount gets error count
-func (hc *HealthChecker) GetErrorCount() int64 {
+func (hc *HealthChecker) GetErrorCount() int {
 	hc.mu.RLock()
 	defer hc.mu.RUnlock()
-	return hc.errorCount
+	return int(hc.errorCount)
 }
 
 // run runs the health check loop
